model enhanced

+++ src/Ecommerce.Domain/Shared/Money.cs
@@ -1,5 +1,7 @@
 ﻿namespace Ecommerce.Domain.Shared;
 
+// Value object representing money with amount and currency
+// Porque Money es inmutable y comparado por valor
 public sealed record Money(decimal Amount, string Currency)
 {
     public Money Add(Money other)
@@ -8,7 +10,7 @@ public sealed record Money(decimal Amount, string Currency)
         return new Money(Amount + other.Amount, Currency);
     }
 
-    public Money Multiply(int factor)
+    public Money Multiply(decimal factor)
     {
         return new Money(Amount * factor, Currency);
     }
